[Refactor] Rank 스케쥴러 분리
